Players origin is now centered.
Class Particle and ParticleGenerator added.

Material has now an BufferedImage attribute.

+++ src/Cam.java
@@ -64,7 +64,7 @@ public class Cam {
         return CamDist;
     }
 
-    public void setCamDist(int camDist) {
+    public void setCamDist(float camDist) {
         CamDist = camDist;
     }
